+++ lib/X1/SA.hs
@@ -7,7 +7,7 @@ module X1.SA ( runSA
              ) where
 
 import Protolude
-import X1.Types.Module
+import X1.Types.Expr1.Module
 import X1.SA.Helpers
 import X1.SA.Types
 import qualified X1.SA.ConflictingTypeAnnDecls as ConflictingTypeAnnDecls
@@ -15,7 +15,7 @@ import qualified X1.SA.ConflictingBindingDecls as ConflictingBindingDecls
 import qualified X1.SA.MissingTopLevelDecls as MissingTopLevelDecls
 
 
-validations :: FilePath -> [Validation [SAError] (Module Decl)]
+validations :: FilePath -> [Validation [SAError] Module]
 validations path =
   map (\f -> f path)
     [ ConflictingBindingDecls.validate
@@ -23,6 +23,6 @@ validations path =
     , MissingTopLevelDecls.validate
     ]
 
-runSA :: FilePath -> Module Decl -> ValidationResult SemanticError
+runSA :: FilePath -> Module -> ValidationResult SemanticError
 runSA path decls = SemanticError <$> analyze (validations path) decls
