+++ lib/X1/Types/Expr1/Impl.hs
@@ -3,7 +3,7 @@ module X1.Types.Expr1.Impl ( Impl(..) ) where
 
 import Protolude
 import X1.Types.Expr1.Pred
-import X1.Types.Expr1 ( Binding )
+import X1.Types.Expr1.Expr ( Binding )
 
 
 data Impl = Impl [Pred] Pred [Binding]
