+++ lib/X1/Types/Expr1/Module.hs
@@ -1,12 +1,12 @@
 
 {-# LANGUAGE DeriveFunctor #-}
 
-module X1.Types.Module ( Decl(..), Module(..) ) where
+module X1.Types.Expr1.Module ( Decl(..), Module(..) ) where
 
 import Protolude hiding ( Fixity(..) )
 import X1.Types.Id
 import X1.Types.Fixity
-import X1.Types.Expr1
+import X1.Types.Expr1.Expr
 import X1.Types.Expr1.ADT
 import X1.Types.Expr1.Trait
 import X1.Types.Expr1.Impl
@@ -21,6 +21,6 @@ data Decl = TypeAnnDecl TypeAnn
           | FixityDecl Fixity Int Id
   deriving (Eq, Show)
 
-newtype Module a = Module [a]
-  deriving (Eq, Show, Functor)
+newtype Module = Module [Decl]
+  deriving (Eq, Show)
